pyproject.toml
[tool.poetry]
name = "skymarket"
version = "0.1.0"
description = ""
authors = ["Vitalii Madiakin <madyakinvitalii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.1"
django-cors-headers = "^3.14.0"
python-dotenv = "^1.0.0"
djoser = "^2.1.0"
django-filter = "^22.1"
django-pytest = "^0.2.0"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"
django-phonenumber-field = "^7.1.0"
phonenumbers = "^8.13.14"
drf-yasg = "^1.21.5"
spectacular = "^0.4.0"
djangorestframework-simplejwt = "^5.2.2"
pytest = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
